Refactor/move query execution (#1246)
* Move QueryProcessor out of model


* Link query_execution


* Resolve conflicts


* Improve codestyle


* Refactor CMakeLists.txt
