+++ irohad/model/CMakeLists.txt
@@ -29,8 +29,8 @@ target_link_libraries(sha3_hash
 add_library(model
     model_crypto_provider_impl.cpp
     impl/model_operators.cpp
-    impl/query_execution.cpp
     )
+
 target_link_libraries(model
     hash
     sha3_hash
