+++ irohad/torii/processor/CMakeLists.txt
@@ -4,7 +4,6 @@ add_library(processors
     )
 
 target_link_libraries(processors PUBLIC
-    model
     rxcpp
     logger
     endpoint
