+++ test/module/iroha-cli/CMakeLists.txt
@@ -20,6 +20,7 @@ target_link_libraries(client_test
     client
     processors
     server_runner
+    query_execution
     )
 target_include_directories(client_test PUBLIC
     ${PROJECT_SOURCE_DIR}/iroha-cli
