+++ test/module/irohad/torii/processor/CMakeLists.txt
@@ -10,4 +10,5 @@ addtest(query_processor_test query_processor_test.cpp)
 target_link_libraries(query_processor_test
     processors
     shared_model_cryptography
+    query_execution
     )
